Replaced prefixes with reference, pluggable topic naming
To centralize the topic name construction, prefix-Strings were replaced
with references to the parent structure (Topology and Project, resp.)
in Project and Topology. This allows to have the logic for naming topics
in a single place. The TopicNameStrategy interface was introduced for,
together with the FullPathStrategy class which replicates the previous
naming convention.

* Removed prefix setters and getters in Topology, Project, and Topic.

* Replaced ArrayList and HashMap with a single LinkedHashMap in Topology.

* Introduced TopicNameStrategy and FullPathStrategy.

* Moved buildNamePrefix logic to FullPathStrategy.

* Changed project getter and setter in Project to retain a copy and to
automatically set reference to parent structure.

* Removed now superfluous updateTopology method from TopologySerde.